Faster hashing without the need for x86 intrinsics

+++ nf/vigor-nat/flowtable.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 
@@ -16,7 +17,7 @@ struct flow {
 struct flowtable;
 
 // Allocates a new flow table. Cannot fail (unless memory allocation fails, in which case the program crashes).
-struct flowtable* flowtable_alloc(uint16_t starting_port, uint64_t expiration_time, uint64_t max_flows);
+struct flowtable* flowtable_alloc(uint16_t starting_port, uint64_t expiration_time, size_t max_flows);
 
 // Gets the external port associated with a given internal flow, at the given time, allocating one if needed. Fails if allocation is required but there is no more space and no flow could be expired.
 bool flowtable_get_internal(struct flowtable* table, uint64_t time, struct flow* flow, uint16_t* out_port);
